requirements.txt
python-telegram-bot[socks]
PySocks
sqlalchemy
python-redmine